requirements.txt
asgiref==3.5.0
Beaker==1.11.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.6
django-cors-headers==3.10.1
django-filter==21.1
django-forestadmin==1.1.4
django-ipware==4.0.2
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
ecdsa==0.17.0
filelock==3.6.0
future==0.18.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-jsonapi==0.24.0
numpy==1.22.3
oic==1.3.0
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.1
platformdirs==2.5.0
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.14.1
pyjwkest==1.4.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2021.3
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
djangorestframework-simplejwt==4.6.0

